Split commit tests
